requirements.txt
pymupdf
sentence-transformers
faiss-cpu
numpy
tqdm
pickle5
transformers
accelerate
torch
matplotlib